Market Design

+++ app/src/main/java/huji/ac/il/stick_defence/BigWoodenTower.java
@@ -6,7 +6,7 @@ import android.content.Context;
  * Created by yahav on 21/08/15.
  */
 public class BigWoodenTower extends Tower{
-    public static final double MAX_HP = 200.0;
+    public static final int MAX_HP = 200;
 
     public BigWoodenTower(Context context, Sprite.Player player){
         super(context, player, R.drawable.big_wooden_tower_blue,
