+++ app/src/main/java/huji/ac/il/stick_defence/StoneTower.java
@@ -6,7 +6,7 @@ import android.content.Context;
  * Created by yahav on 21/08/15.
  */
 public class StoneTower extends Tower{
-    public static final double MAX_HP = 300.0;
+    public static final int MAX_HP = 300;
 
     public StoneTower(Context context, Sprite.Player player){
         super(context, player, R.drawable.stone_tower_blue,
